Fix _binary_array_to_hex.
This is a fix for the given problems:

1. When a hash is generated with hash_size=6 _binary_array_to_hex
will give a 8 digit hexadecimal representation for it while the
correct would be a 9 digit hecadecimal representation. Retrieving
a hash from the hex value in these cases is impossible.

2. _binary_array_to_hex gives no output when hash_size=2.

3. The _binary-to-hexadecimal conversion is not interoperable with
other implementations of the same algorithm.

Fixing _binary_array_to_hex demands hex_to_hash to be changed as well.
That makes hex_to_hash incompatible with hex values generate with
previous versions of ImageHash. For that reason, the old
implementation is available as old_hex_to_hash in order to allow
users convert old hex values back to hashes.

Regarding the new version of hex_to_hash:
- It assumes all hashes are bidimensional arrays with dimensions
hash_size * hash_size. This is the current behavior of all hashes.
- The implementation does not work for hash_size < 2.